handlers refactor
